# src/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server.",
        "title": "Fiber Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/CreateFact/": {
            "post": {
                "description": "create fact.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "createFact"
                ],
                "summary": "create fact.",
                "parameters": [
                    {
                        "description": "request",
                        "name": "session",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateFactRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateFactResponse"
                        }
                    }
                }
            }
        },
        "/GetFact/": {
            "post": {
                "description": "get fact.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "getFact"
                ],
                "summary": "get fact.",
                "parameters": [
                    {
                        "description": "request",
                        "name": "session",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.GetFactRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetFactResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CreateFactRequest": {
            "type": "object",
            "properties": {
                "facts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Fact"
                    }
                }
            }
        },
        "dto.CreateFactResponse": {
            "type": "object"
        },
        "dto.GetFactRequest": {
            "type": "object",
            "properties": {
                "indicator_to_mo_id": {
                    "type": "string"
                },
                "period_end": {
                    "type": "string"
                },
                "period_key": {
                    "type": "string"
                },
                "period_start": {
                    "type": "string"
                }
            }
        },
        "dto.GetFactResponse": {
            "type": "object",
            "properties": {
                "indicator_to_mo_id": {
                    "type": "string"
                },
                "period_end": {
                    "type": "string"
                },
                "period_key": {
                    "type": "string"
                },
                "period_start": {
                    "type": "string"
                }
            }
        },
        "models.Fact": {
            "type": "object",
            "properties": {
                "auth_user_id": {
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                },
                "fact_time": {
                    "type": "string"
                },
                "indicator_to_mo_fact_id": {
                    "type": "string"
                },
                "indicator_to_mo_id": {
                    "type": "string"
                },
                "is_plan": {
                    "type": "string"
                },
                "period_end": {
                    "type": "string"
                },
                "period_key": {
                    "type": "string"
                },
                "period_start": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "JWT": {
            "description": "JWT security accessToken. Please add it in the format \"Bearer {AccessToken}\" to authorize your requests.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}